refactor(dashboard-widget-container): separate internal logics to composables
